update logic code voucher
