+++ chillStore/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:sqlserver://localhost:1433;databaseName=group4_swp_chillStore_Final10;encrypt=true;trustServerCertificate=true
 spring.datasource.username=sa
-spring.datasource.password=080404
+spring.datasource.password=123456789
 spring.datasource.driver-class-name=com.microsoft.sqlserver.jdbc.SQLServerDriver
 
 spring.jpa.hibernate.ddl-auto=validate
